libfreerdp-core: move timezone code from info.c to timezone.c
